refactor: move strategy fetching into trade-executor/client
