+++ src/lib/trade-executor/helpers/chart.ts
